+++ src/api/Api.jsx
@@ -67,6 +67,17 @@ export async function apiCallToGetCarbonThresholds() {
   return resData;
 }
 
+export async function apiCallToGetEntityTemplate(entity) {
+  const response = await fetch("http://localhost:8080/" + entity + "/template");
+  const resData = await response.json();
+
+  if (!response.ok) {
+    throw new Error("Failed to get entity template.");
+  }
+
+  return resData;
+}
+
 export async function apiCallToGetSingleEntity(id, name, entity) {
   let endpoint;