+++ src/components/Main/Company/Table/RowDetail/DetailView/DetailView.jsx
@@ -27,7 +27,6 @@ import {
 const DetailView = memo(function DetailView({
   entity,
   entityName,
-  fieldName,
   updateRowHandler,
   deleteRowHandler,
   rowIndex,