+++ src/components/Main/Company/Table/RowDetail/DetailView/DetailView.utils.js
@@ -17,7 +17,6 @@ export function determineUniqueFieldName(entityName) {
 
 function determinateFieldType(field) {
   const numberFields = [
-    "id",
     "apartmentNumber",
     "capacity",
     "carbonLimit",